# Optimize Chunk traverse