I noticed that `Stream.evalMapChunk`, which is a common operation for me, goes through `Chunk.traverse`, however that is implemented in a pretty expensive way: via foldRight which wraps everything in another layer of Eval.

I instead implemented a tree-based approach (to keep stack safety) but without wrapping in another Eval layer.