@johnynek 2020-07-11
Thanks for the comment @SystemFw .

One difference here is that it won't blow the stack on very big chunks for non-trampolining `F[_]`. In practice, that might be rare for Stream since generally people are using an effect type that is trampolined, but still, for a few folks that may notice.

For anyone using Chunk on its own, they might really care that they can traverse them for sizes up to 100k, and non-trampolined `F[_]`.

I didn't benchmark this, but my guess is that if you have an `evalMapChunk` that is not doing a "real" IO (e.g. not doing network or file access), maybe something that is just catching the possibility of exception, then the difference will be pretty large (maybe 2x or more?) since you don't have do use both the Eval and the IO evaluation.