@SystemFw 2020-07-11
The old traverse will also be stack safe (just tried traversing with Option on a Chunk with 1 million elems), but I'm okay with merging if we're confident there is going to be a perf benefit :)