@johnynek 2020-07-11
I'll make a benchmark before we merge. Let's be rigorous.