@johnynek 2020-07-11
Option may short circuit on None so unless you have no Nones I'm not sure that is a real test.