@SystemFw 2020-07-11
This is what I've used:
```
 val source = (0 to 1000000).toVector
 val tt = Chunk.vector(source).traverse(_.pure[Option])
```