@johnynek 2020-07-11
On main:
```
scala> Chunk.seq(0 until 1000000).traverse { i => () => i }
val res0: () => fs2.Chunk[Int] = cats.instances.Function0Instances$$anon$4$$Lambda$6313/809375313@287dcd6

scala> res0()
java.lang.StackOverflowError
  ... 1024 elided
```

with this change:
```
scala> Chunk.seq(0 until 1000000).traverse { i => () => i }
val res0: () => fs2.Chunk[Int] = cats.instances.Function0Instances$$anon$4$$Lambda$7732/324635468@3467ff7a

scala> res0()
val res1: fs2.Chunk[Int] = Chunk(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100, 101, 102, 103, 104, 105, 106, 107, 108, 109, 110, 111, 112, 113, 114, 115, 116, 117, 118, 119, 120, 121, 122, 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 136, 137, 138, 139, 140, 141, 142, 143, 144, 145, 146, 147, 148, 149, 150, 151, 152, 153, 154, 155, 156, 157, 158, 159, 160, 161, 162, 163, 164, 165, 166, 167, 168, 169...
```

Not sure the difference, but the `Applicative[F]` is only seeing log depth in my code so even a highly unsafe monad like `Function0` works.