@johnynek 2020-07-11
I was wrong. This is about the same speed (maybe modestly faster), the win is the stack safety alone it looks like.

Also, this design has N log N calls to F.map or F.map2 because it builds a tree up. The original implementation does use Eval, but it only calls F.map2 N times (in a linear chain). So, this PR is a bit behind the 8-ball due to that.

I think it is worth merging for the added stack safety, but the performance motivation is not there.

If we don't care about stack safety (which we don't seem to have in all cases currently), I bet you can go faster with a linear chain build without Eval and foldRight: just iterate through the list right to left building up an inner `List` then 1 final `.map(Chunk.list)` at the end.

Here are benchmark results:
```
on main:

[info] Benchmark                (chunkCount)  (chunkSize)   Mode  Cnt       Score        Error  Units
[info] ChunkBenchmark.traverse             1           16  thrpt    6  881546.345 ± 140421.631  ops/s
[info] ChunkBenchmark.traverse             1          256  thrpt    6   44430.572 ±   2828.650  ops/s
[info] ChunkBenchmark.traverse            20           16  thrpt    6   36710.054 ±    217.760  ops/s
[info] ChunkBenchmark.traverse            50           16  thrpt    6   15346.930 ±     38.888  ops/s
[info] ChunkBenchmark.traverse           100           16  thrpt    6    7039.792 ±    153.096  ops/s
// the ones omitted had stack overflows

on this PR:
[info] Benchmark                (chunkCount)  (chunkSize)   Mode  Cnt       Score      Error  Units
[info] ChunkBenchmark.traverse             1           16  thrpt    6  852749.128 ± 5820.295  ops/s
[info] ChunkBenchmark.traverse             1          256  thrpt    6   48765.634 ±  403.656  ops/s
[info] ChunkBenchmark.traverse            20           16  thrpt    6   39116.667 ±  203.117  ops/s
[info] ChunkBenchmark.traverse            50           16  thrpt    6   14817.411 ±  385.322  ops/s
[info] ChunkBenchmark.traverse           100           16  thrpt    6    7027.628 ±  944.514  ops/s

// these complete on this PR, but not on main:

[info] ChunkBenchmark.traverse             1         4096  thrpt    6    2944.750 ±   10.933  ops/s
[info] ChunkBenchmark.traverse            20          256  thrpt    6    2322.770 ±   65.179  ops/s
[info] ChunkBenchmark.traverse            20         4096  thrpt    6      89.735 ±   17.816  ops/s
[info] ChunkBenchmark.traverse            50          256  thrpt    6     816.425 ±   91.386  ops/s
[info] ChunkBenchmark.traverse            50         4096  thrpt    6      30.275 ±    1.905  ops/s
[info] ChunkBenchmark.traverse           100          256  thrpt    6     228.106 ±   14.465  ops/s
[info] ChunkBenchmark.traverse           100         4096  thrpt    6      12.225 ±    1.451  ops/s
```