@johnynek 2020-07-14
It occurred to me in a comment here:

https://github.com/typelevel/cats/issues/3517#issuecomment-657916049

the magic here is the tree vs the linear structure, not the fact that it is a binary tree. It is a bit more complex to implement, but a fan-out of say 128 children per node or even 256 will preserve the log growth in stack size but will reduce the overhead of building the tree since we will build up linear chains of 128 or 256.

This may allow us to actually see a performance boost in addition to the stack safety achieved.